Package changes, Arbitrator and Behaviors
Added new "RobotBrain" class for centre of operations, where all sensors and actuators can be initialised. Other packages can then import this to use sensors etc. 

Set up behavior templates, need to add actions and priorities of behaviors, but thought useful to have a template. 

Comments on whether or not most methods can be stored in "RobotBrain", or if we should have new classes for sets of related operations?