Color Sensor Localisation
Modified the robot class to fit the new build and marked irrelevant
classes as deprecated to show that they’re not to be used before. Was
halted with “could not open port” error as I had not properly removed
touch sensor classes from the robot class. Am now working on getting
localisation behaviour from the two colour sensors working. Having to
reconfigure the color sensor ranges as the color sensors are higher up.
Color Sensor monitor will need to be improved to distinguish colors
more efficiently.